requirements.txt
google-genai>=1.0.0
resend>=2.0.0
requests>=2.32.0
